requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
mcp>=1.0.0
uvicorn>=0.32.1
PyQt5>=5.15.0
httpx>=0.25.0